move BlueBG to Common
