+++ src/components/Speciality/talks.js
@@ -5,7 +5,7 @@ import { Row, Col, Grid } from '../grid'
 import { SectionTitle } from '../Typography'
 import { Padding } from 'styled-components-spacing'
 import StyledLink from '../Common/StyledLink'
-import BlueBackground from '../BlueBG'
+import BlueBackground from '../Common/BlueBG'
 import styled from 'styled-components'
 import CompactVideoLink from '../Common/CompactVideoLink'
 import theme from '../../utils/theme'
