+++ src/components/Training/Approach.js
@@ -4,7 +4,7 @@ import { Padding } from 'styled-components-spacing'
 import { SectionTitle, Subtitle, BodyPrimary } from '../Typography'
 import { Grid, Row, Col } from '../grid'
 import Hr from '../Common/Hr'
-import BlueBackground from '../BlueBG'
+import BlueBackground from '../Common/BlueBG'
 import BackgroundGraphic from '../Common/BackgroundGraphic'
 import TitleAndList from '../Common/TitleAndList'
 import theme from '../../utils/theme'
