+++ src/components/layout.js
@@ -5,7 +5,7 @@ import { StaticQuery, graphql } from 'gatsby'
 import { Location } from '@reach/router'
 import Header from './Header'
 import './layout.css'
-import BlueBackground from './BlueBG'
+import BlueBackground from './Common/BlueBG'
 import theme from '../utils/theme'
 import GlobalStyle from '../utils/globalStyle'
 import Footer from './Footer'
