+++ stories/compactVideoLink.stories.js
@@ -3,7 +3,7 @@ import { storiesOf, addDecorator } from '@storybook/react'
 import Theme from './theme'
 import CompactVideoLink from '../src/components/Common/CompactVideoLink'
 import { Row } from '../src/components/grid'
-import BlueBackground from '../src/components/BlueBG'
+import BlueBackground from '../src/components/Common/BlueBG'
 import GreyBackground from '../src/components/GreyBG'
 
 addDecorator(Theme)
