+++ stories/standaloneVideoLink.stories.js
@@ -6,7 +6,7 @@ import theme from '../src/utils/theme'
 import GlobalStyle from '../src/utils/globalStyle'
 import StandaloneVideoLink from '../src/components/Common/StandaloneVideoLink'
 import { Row } from '../src/components/grid'
-import BlueBackground from '../src/components/BlueBG'
+import BlueBackground from '../src/components/Common/BlueBG'
 import GreyBackground from '../src/components/GreyBG'
 
 const Theme = storyFn => (
